backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
firebase-admin
requests
openai
langchain
langchain-openai
langgraph
